rename: balanceGame rushGame 으로 명칭 변경
