+++ client/src/features/Rush/RushGame/Background.tsx
