+++ client/src/features/Rush/RushGame/RushGameCard/RushCardCurrentRatio.tsx
