+++ client/src/features/Rush/RushGame/RushGameCard/RushCardResultDescription.tsx
