+++ client/src/features/Rush/RushGame/RushGameSection/CardOptions.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from "react";
 import { motion } from "framer-motion";
 import { ASCEND, DISSOLVE, SCROLL_MOTION } from "@/constants/animation.ts";
-import RushCardComparison from "@/features/Rush/BalanceGame/BalanceGameCard/RushCardComparison.tsx";
-import RushCountDown from "@/features/Rush/BalanceGame/BalanceGameCard/RushCountDown.tsx";
+import RushCardComparison from "@/features/Rush/RushGame/RushGameCard/RushCardComparison.tsx";
+import RushCountDown from "@/features/Rush/RushGame/RushGameCard/RushCountDown.tsx";
 import useCountDown from "@/hooks/useCountDown.ts";
 
 export default function CardOptions() {
