+++ client/src/features/Rush/RushGame/RushGameSection/FinalResult.tsx
