+++ client/src/pages/Rush/index.tsx
@@ -1,9 +1,8 @@
 import Footer from "@/components/Footer";
 import Notice from "@/components/Notice";
 import { RUSH_SECTIONS } from "@/constants/PageSections/sections.ts";
-import { BalanceGameProvider } from "@/contexts/balanceGameContext.tsx";
+import { RushGameProvider } from "@/contexts/rushGameContext.tsx";
 import {
-    BalanceGame,
     CasperCharge,
     CasperComfortable,
     CasperFar,
@@ -16,6 +15,7 @@ import {
     Intro,
     ReasonFirst,
     ReasonSecond,
+    RushGame,
 } from "@/features/Rush";
 import useHeaderStyleObserver from "@/hooks/useHeaderStyleObserver.ts";
 import useScrollTop from "@/hooks/useScrollTop.tsx";
@@ -28,9 +28,9 @@ export default function Rush() {
 
     return (
         <div ref={containerRef} className="h-screen overflow-auto snap-y snap-mandatory">
-            <BalanceGameProvider>
-                <BalanceGame id={RUSH_SECTIONS.BALANCE_GAME} />
-            </BalanceGameProvider>
+            <RushGameProvider>
+                <RushGame id={RUSH_SECTIONS.RUSH_GAME} />
+            </RushGameProvider>
             <Intro id={RUSH_SECTIONS.INTRO} />
             <FAQ id={RUSH_SECTIONS.FAQ} />
             <ElectricReason id={RUSH_SECTIONS.ELECTRIC_REASON} />
