+++ client/src/types/rushGame.ts
@@ -1,11 +1,11 @@
 import React from "react";
 import { GetRushUserParticipationStatusResponse } from "@/types/rushApi.ts";
 
-export interface BalanceGameContextType {
+export interface RushGameContextType {
     gameState: {
         phase: "PRE_EVENT" | "EVENT_RUNNING" | "EVENT_ENDED";
         userParticipated: boolean;
     };
-    setGameState: React.Dispatch<React.SetStateAction<BalanceGameContextType["gameState"]>>;
+    setGameState: React.Dispatch<React.SetStateAction<RushGameContextType["gameState"]>>;
     updateUserParticipationStatus: (response: GetRushUserParticipationStatusResponse) => void;
 }
